% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-create_alpha_table.R
\name{create_alpha_table}
\alias{create_alpha_table}
\title{Creates a table of alpha diversity metrics per sample. Optionally saves the table as a .tsv file.}
\usage{
create_alpha_table(phyloseq_object, outputpath = FALSE)
}
\arguments{
\item{phyloseq_object}{The phyloseq object containing the target dataset.}

\item{outputpath}{(optional) A filepath to the folder that the table should be saved into.}
}
\value{
A table of alpha diversity metrics per sample.
}
\description{
Creates a table of alpha diversity metrics per sample. Optionally saves the table as a .tsv file.
}
