% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-create_rarefaction_curve.R
\name{create_rarefaction_curve}
\alias{create_rarefaction_curve}
\title{Creates a rarefaction curve using vegan's rarecurve().}
\usage{
create_rarefaction_curve(phyloseq_object, step_size = 20)
}
\arguments{
\item{phyloseq_object}{The phyloseq object to be loaded.}

\item{step_size}{The intervals the rarefaction curve should be crafted on.}
}
\value{
A rarefaction curve.
}
\description{
Creates a rarefaction curve using vegan's rarecurve().
}
\seealso{
\url{https://www.rdocumentation.org/packages/vegan/versions/2.4-2/topics/rarefy}
}
