% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-create_tree.R
\name{create_tree}
\alias{create_tree}
\title{Creates a phylogenetic tree using phangorn and DECEIPHER.}
\usage{
create_tree(seqtab)
}
\arguments{
\item{seqtab}{A table of sequences. Can be generated via the \code{\link[=DADA2_main]{DADA2_main()}} function.}
}
\value{
A phylogenetic tree object, used to make a full phyloseq object.
}
\description{
Creates a phylogenetic tree using phangorn and DECEIPHER.
}
\seealso{
\code{\link[=DADA2_main]{DADA2_main()}} for sequence table creation.
}
