% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-DADA2_main.R
\name{DADA2_main}
\alias{DADA2_main}
\title{Contains all DADA2 preprocessing, denoising, merging, and chimera removal processes.}
\usage{
DADA2_main(
  fastq_folderpath,
  outputpath,
  trim_left,
  truncF,
  truncR,
  forwardReadPattern,
  reverseReadPattern,
  max_EE_allowed = c(2, 2),
  max_N_allowed = 0,
  max_length_allowed = 10000,
  min_length_allowed = 0,
  min_quality_allowed = 0,
  remove_phiX = TRUE,
  verbose_output = TRUE,
  pooling = FALSE
)
}
\arguments{
\item{fastq_folderpath}{The filepath to the directory containing the FASTQ files to process.}

\item{outputpath}{The filepath to the directory where all filtered sequences and summary data should be output.}

\item{trim_left}{The number of base pairs to trim from the start of all sequences (min 0).}

\item{truncF}{The base number that forward reads should truncate at.}

\item{truncR}{The base number that reverse reads should truncate at.}

\item{forwardReadPattern}{A filename identifier that identifies FASTQ files with forward reads (ex: "_F").}

\item{reverseReadPattern}{A filename identifier that identifies FASTQ files with reverse reads (ex: "_R").}

\item{max_EE_allowed}{A vector containing the maximum error rates allowed (default c(2,2)).}

\item{max_N_allowed}{The maximum amount of unknown bases labelled N to allow before discarding a read (default 0).}

\item{min_length_allowed}{The minimum length a read can be. If the read length < min_length_allowed, the read is discarded (default 0).}

\item{min_quality_allowed}{The minimum quality score that a base can have before the entire read is discarded (default 0).}

\item{remove_phiX}{A boolean signifying whether reads assigned to phiX should be removed (default TRUE).}

\item{verbose_output}{A boolean signifying whether the main DADA2 step should be verbose or silent (default TRUE).}

\item{pooling}{A boolean signifying whether DADA2 should pool samples together prior to sample inference (default FALSE).}
}
\value{
Returns a table of sequences, all of which are non-chimeric and have been denoised and merged.
}
\description{
Contains all DADA2 preprocessing, denoising, merging, and chimera removal processes.
}
