% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-filepath_management.R
\name{fileplatform}
\alias{fileplatform}
\title{Helper function to replace slashes and backslashes depending on the OS type.}
\usage{
fileplatform(pathstring)
}
\arguments{
\item{pathstring}{A string of a filepath containing slashes or backslashes.}
}
\value{
A OS-dependent filepath string.
}
\description{
Helper function to replace slashes and backslashes depending on the OS type.
}
\keyword{internal}
