% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-format_q2_metadata.R
\name{format_q2_metadata}
\alias{format_q2_metadata}
\title{Formats metadata from a QIIME2-compatible .tsv file to a data frame, and prepares it for downstream use.}
\usage{
format_q2_metadata(metadata_path)
}
\arguments{
\item{metadata_path}{A filepath to the metadata file.}
}
\value{
A dataframe of the imported metadata.
}
\description{
Formats metadata from a QIIME2-compatible .tsv file to a data frame, and prepares it for downstream use.
}
