% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-handoff_to_phyloseq.R
\name{handoff_to_phyloseq}
\alias{handoff_to_phyloseq}
\title{Creates a phyloseq object from the supplied data. If outputpath is specified, save the phyloseq object.}
\usage{
handoff_to_phyloseq(
  sequence_table,
  metadata,
  taxa_table,
  tree = FALSE,
  outputpath = FALSE,
  substitute_ASV_names = FALSE
)
}
\arguments{
\item{sequence_table}{A sequence table to include. Can be generated from \code{\link[=DADA2_main]{DADA2_main()}}.}

\item{metadata}{A metadata table to include. Can be generated from \code{\link[=format_q2_metadata]{format_q2_metadata()}}.}

\item{taxa_table}{A taxonomy table to include. Can be generated from \code{\link[=label_taxonomy]{label_taxonomy()}}.}

\item{tree}{(optional) A phylogenetic tree to include. Can be generated from \code{\link[=create_tree]{create_tree()}}.}

\item{outputpath}{(optional) A filepath to a folder to output the saved phyloseq object as an RDS object.}

\item{substitute_ASV_names}{(optional) A boolean specifying whether ASV names should be substituted with ASV1, ASV2... The names will be stored in "dna".}
}
\value{
A phyloseq object consisting of the supplied sequence table, metadata, taxonomy table, and (optional) phylogenetic tree.
}
\description{
Creates a phyloseq object from the supplied data. If outputpath is specified, save the phyloseq object.
}
\seealso{
\code{\link[=DADA2_main]{DADA2_main()}} for sequence table creation, \code{\link[=format_q2_metadata]{format_q2_metadata()}} for metadata importing, \code{\link[=label_taxonomy]{label_taxonomy()}} for taxonomy table creation, and \code{\link[=create_tree]{create_tree()}} for phylogenetic tree creation.
}
