% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-label_taxonomy.R
\name{label_taxonomy}
\alias{label_taxonomy}
\title{Assigns taxonomy classifications to a provided table of sequences.}
\usage{
label_taxonomy(seqtab, pathToDatabase, species = FALSE, pathToSpecies = FALSE)
}
\arguments{
\item{seqtab}{The provided table of sequences. Can be generated from DADA2_main().}

\item{pathToDatabase}{A filepath to a genus-level classification database compatible with DADA2.}

\item{species}{A boolean specifying whether specie-level classification should be attempted.}

\item{pathToSpecies}{A filepath to a species-level classification database compatible with DADA2.}
}
\value{
Returns a table of classified sequences.
}
\description{
Assigns taxonomy classifications to a provided table of sequences.
}
