% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-load_phyloseq.R
\name{loadphyloseq}
\alias{loadphyloseq}
\title{A helper function to quickly load phyloseq objects. Currently unused.}
\usage{
loadphyloseq(output_dir, rarefy_samples = FALSE)
}
\arguments{
\item{output_dir}{A filepath to the output directory, without the final slash.}

\item{rarefy_samples}{A boolean showing whether a rarefied phyloseq object should be loaded or not.}
}
\value{
The loaded phyloseq object.
}
\description{
A helper function to quickly load phyloseq objects. Currently unused.
}
\keyword{internal}
