% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-plot_alpha_all.R
\name{plot_alpha_all}
\alias{plot_alpha_all}
\title{Plots all general alpha diversity metrics on one plot.}
\usage{
plot_alpha_all(phyloseq_object)
}
\arguments{
\item{phyloseq_object}{The phyloseq object containing the data to be visualized.}
}
\value{
A plot showing all general metrics of alpha diversity.
}
\description{
Plots all general alpha diversity metrics on one plot.
}
