% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-plot_alpha_group.R
\name{plot_alpha_group}
\alias{plot_alpha_group}
\title{Creates a list of alpha diversity plots with samples separated into groups based on the provided grouping variable.}
\usage{
plot_alpha_group(
  phyloseq_object,
  grouping_variable,
  alpha_types = c("Observed", "Chao1", "ACE", "Shannon", "Simpson", "InvSimpson",
    "Fisher")
)
}
\arguments{
\item{phyloseq_object}{The phyloseq object containing the dataset to be plotted.}

\item{grouping_variable}{The column name in the metadata that the samples should be grouped by. Multiple columns can be looped over if provided in a vector.}

\item{alpha_types}{The types of alpha diversity metrics to plot.}
}
\value{
Returns a list of alpha diversity plots (if grouping variable isn't a vector) or a list of lists (if grouping variable is a vector). Raises an error if alpha_types isn't a vector
}
\description{
Creates a list of alpha diversity plots with samples separated into groups based on the provided grouping variable.
}
