% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-rarefy_phyloseq.R
\name{rarefy_phyloseq}
\alias{rarefy_phyloseq}
\title{Rarefies samples in the provided phyloseq object to the specified rarefaction depth. Optionally saves the resulting phyloseq object.}
\usage{
rarefy_phyloseq(phyloseq_object, rarefaction_depth, outputpath = FALSE)
}
\arguments{
\item{phyloseq_object}{The phyloseq object to rarefy.}

\item{rarefaction_depth}{A number specifying the number of reads to rarefy to per sample.}

\item{outputpath}{(optional) A filepath specifying the folder to save the rarefied object to.}
}
\value{
A rarefied phyloseq object.
}
\description{
Rarefies samples in the provided phyloseq object to the specified rarefaction depth. Optionally saves the resulting phyloseq object.
}
